# [SPARK-34119][SQL] Keep necessary stats after partition pruning