@peter-toth 2021-01-16
This looks much better than my PR: https://github.com/apache/spark/pull/29132
@wangyum I wonder if you checked other than q14 query improvements?