@cloud-fan 2021-01-18
This patch alone doesn't change the query plan, so I think it's fine to skip benchmark as it shouldn't change anything.