@cloud-fan 2021-01-18
shall we do the same in PruneHiveTablePartitions?