@wangyum 2021-01-18
> shall we do the same in PruneHiveTablePartitions?

Yes, PruneHiveTablePartitions also has this issue.